added operator pack

+++ src/pstream.c
@@ -1233,8 +1233,8 @@ void pstreamCheckDatarange(pstream_t *pstreamptr, int varID, double *array, int 
       if ( datatype == DATATYPE_INT8  || datatype == DATATYPE_UINT8 ||
 	   datatype == DATATYPE_INT16 || datatype == DATATYPE_UINT16 )
 	{
-	  smin = NINTD(smin);
-	  smax = NINTD(smax);
+	  smin = (int) NINTD(smin);
+	  smax = (int) NINTD(smax);
 	}
 
       if      ( datatype == DATATYPE_INT8   ) { vmin =        -128.; vmax =        127.; }
